rearrange and reorganize. more logical separation of functionality
RPiGPIO.class.php: Commented code a bit. Removed default GPIO pin number
and added some code to loosely enforce for valid GPIO pin numbers. This
appears to work well now, though unit testing would be good.

rfid_key.class.php: Added author and related comments.  Removed all
comments and code that were not specific to basic RFID-key parsing and
handling. In other words, all non-rfid stuff that might be related to a
given rfid-key, was removed to make this concept clean and simple.
People now have a RFIDs rather than the other way around.

BWRPi_RFIDkeyd.php had too much going on.  I moved its contents to the
new files: FrontDoor.php, RFIDResource.php, bwdb.php, and main.php.

RFIDResource.php is a keeper of code that relates to abstract resources
that may be unlocked by an RFID-key. this is in prepartion for access to
things other than the front door.

FrontDoor.php contains information about a specific resource, the front
door of the makerspace.  It depends on, extends and implements
interfaces found in RFIDResource.php.  FrontDoor also contains a
"Contoller" element that ties the RFID reader to the GPIO door-lock
element.  This is a central piece.

main.php is the starting point for the daemon. This file is almost
empty so far, and may stay that way.

bwdb.php shows some sketches for different ways of modelling the data
and the interface to the database.  This file may disappear and much of
its contents will be thrown away.  I'm still deciding what to do, and
this file is not presently used by any other.

main.php and associated code works presently on a Rasp-Pi.  It starts
and waits for input from the RFID reader on the UART. GPIO pin #2 (on
RPi v2) is the pin used to for unlocking the door.  Presently, when an
RFID card is read and parsed successfully (checksum is good) then the
GPIO is brought high for 10 seconds after which is returns to its
ordinary wait (locked) state. Thus we can use this in its present state
to unlock a door, though it doesn't differentiate bewtween RFID keys.