Fixes some bugs and makes some needed improvements. This code is used.
Since we installed this on the front door a number of bug fixes were
made and some important improvements were added. I don't have enough
time now to talk about all the changes. This code is working well
though.

BUG: The system has a hicup every so often.  Occasionally the serial
line interface to the RFID reader enters a bad states where nothing
sensible can be read from it.  It appears to happen after a long(er)
period of front-door silence (no one using the door).  Right now I
handle this case by exiting; which is a tolerable fix because the parent
script will capture the exit event and restart the daemon after a n
second delay (where 'n' is 1 or so).  When this happens the frontdoor
can fail to open 2 or 3 times before successfully opening the door.
It's functional.

I'm leaving a lot out of this description.  Please look at the diff.  I
will probably write more later.  I added some simple installation
instructions which likely leave much to be desired, but are accurate.